requirements.txt
python-debian
requests